+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Changes
 
-* `@apostrophecms/migration:after` handler now runs the migration requirements like `insertIfMissing`, `implementParkAllInDefaultLocale`, `replicate` and `implementParkAllInOtherLocales`.
+* `@apostrophecms/migration:requirements` handler now runs the migration requirements like `insertIfMissing`, `implementParkAllInDefaultLocale`, `replicate` and `implementParkAllInOtherLocales`.
 
 ### Fixes